work with films controller, upload files

+++ ex00/Cinema/src/main/java/edu/school21/cinema/repositories/MovieRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface MovieRepository {
     public void addMovie(Movie movie);
     public void updateMovie(Movie movie);
-    public void removeMovie(int id);
-    public Movie getMovieById(int id);
+    public void removeMovie(Long id);
+    public Movie getMovieById(Long id);
     public List<Movie> listMovies();
 }
